requirements-azure.txt
cat > requirements-azure.txt << 'EOF'
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-functions==1.17.0
azure-identity==1.17.0
azure-cosmos==4.4.0
azure-sqlalchemy==1.4.52
sqlalchemy[asyncio]==2.0.23
pydantic-settings==2.1.0
alembic==1.13.1
structlog==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
EOF
